+++ Medanit-backend/routes/comments.js
@@ -1,7 +1,7 @@
 const debug = require('debug')('app:comments');  // set/export DEBUG=app:comments
 const express = require('express');
 
-const { Comment, validatePost, validatePut } = require('../Models/comment');
+const { Comment, validatePost, validatePut, validateId } = require('../Models/comment');
 const { Post } = require('../Models/post');
 
 const router = express.Router({mergeParams:true});
@@ -15,20 +15,30 @@ router.get('/', async (req, res) => {
     const skipFactor = pageNumber ? (pageNumber - 1) * pageSize : 0;
     const limitFactor = pageNumber ? pageSize : 0;
 
-    debug(`Page size: ${pageSize}\nPage number: ${pageNumber}`);
+    // debug(`Page size: ${pageSize}\nPage number: ${pageNumber}`);
 
-    // check if the post exists
-    const post = await Post.findById(postId);
-    if(!post) return res.status(404).send('Post with provided Id does not exist!');
-
-    const comments = await Comment
-        .find( { post_id: postId } )
-        .skip( skipFactor )
-        .limit( limitFactor )
-        .sort( { date: 1 } )
-        .select()
+    if(!validateId(postId)) return res.status(404).send('Invalid Post Id!');
     
-    res.send(comments);
+    try{
+        // check if the post exists
+        const post = await Post.findById(postId);
+        if(!post) return res.status(404).send('Post with provided Id does not exist!');
+
+        let comments = await Comment
+            .find( { post_id: postId } )
+            .skip( skipFactor )
+            .limit( limitFactor )
+            .sort( { date: -1 } )
+            .select()
+        
+        comments = updateDate(comments); 
+
+        res.send(comments);
+
+    }catch(err){
+        debug(err.message);
+        res.status(500).send("Internal server error while getting comments!");
+    }
 
 });
 
@@ -40,65 +50,88 @@ router.post('/', async (req, res) => {
     const { error, result } = validatePost(req.body);
     if (error) return res.status(400).send(error.message);
 
+    if(!validateId(postId)) return res.status(404).send('Invalid Post Id!');
+
     // check if the post exists
-    const post = await Post.findById(postId);
-    if(!post) return res.status(404).send('Post with provided Id does not exist!');
+    try{
+        const post = await Post.findById(postId);
+        if(!post) return res.status(404).send('Post with provided Id does not exist!');
 
-    let comment = new Comment({
-        user_id: req.body.user_id,
-        content: req.body.content,
-        post_id: postId,
-        date: new Date()
-    });
+        let comment = new Comment({
+            user_id: req.body.user_id,
+            content: req.body.content,
+            post_id: postId,
+            date: new Date()
+        });
 
-    comment = await comment.save()
-
-    res.status(201).send(comment);
+        comment = await comment.save()
+        res.status(201).send({...comment._doc, date: comment.date});
+        
+    }catch(err){
+        debug(err.message);
+        res.status(500).send("Internal server error while saving comment!")
+    }
 });
 
 /* PUT METHOD COMMENT */
 router.put('/:id', async (req, res) => {
+    const postId = req.params.post_id;
     const id = req.params.id;
 
     const { error, value } = validatePut(req.body);
     if (error) return res.status(400).send(error.message);
 
-    // check if it exists first
-    const oldComment = await Comment.findById(id);
-    if(!oldComment) return res.status(404).send('Resource not found!');
+    if(!validateId(id)) return res.status(404).send('Invalid Comment Id!');
+    if(!validateId(postId)) return res.status(404).send('Invalid Post Id!');
+
+
     
-    let result;
+    
+    try{
+        // check if it exists first
+        const oldComment = await Comment.findById(id);
+        if(!oldComment) return res.status(404).send('Resource not found!');
 
-    if(req.query.action){
+        let result;
 
-        handleUpvoteDownvoteNotification(req);
-        result = await handleUpvoteDownvoteDB(req);
+        if(req.query.action){
+            handleUpvoteDownvoteNotification(req);
+            result = await handleUpvoteDownvoteDB(req);
+        }else{
+            let newComment = {
+                content: req.body.content || oldComment.content
+            };
+            debug(newComment);
+            oldComment.set(newComment);
+            result = await oldComment.save();
+        }
 
-    }else{
-        
-        let newComment = {
-            content: req.body.content || oldComment.content
-        };
+        res.status(204).send(result);
 
-        debug(newComment);
-
-        oldComment.set(newComment);
-        result = await oldComment.save();
-
+    }catch(err){
+        debug(err.message);
+        res.status(500).send("Internal server error while modifying comment!")
     }
-
-
-    res.status(204).send(result);
 });
 
 
 /* DELETE COMMENT */
 router.delete('/:id', async (req, res) => {
     const id = req.params.id;
+    const postId = req.params.post_id;
 
-    const result = await Comment.deleteOne( { _id: id } )
-    if(!result) res.status(404).send('Resource not found!');
-    res.send(result);
+    if(!validateId(postId)) return res.status(404).send('Invalid Post Id!');
+    if(!validateId(id)) return res.status(404).send('Invalid Comment Id!');
+    
+    try{
+        const result = await Comment.deleteOne( { _id: id } )
+        if(!result) res.status(404).send('Resource not found!');
+        res.send(result);
+    }catch(err){
+        debug(err);
+        res.status(500).send("Internal server error while trying to delete comment!");
+    }
+    
 });
 
 
@@ -139,6 +172,18 @@ async function handleUpvoteDownvoteDB(req){
             { new: true });
     }
 
+}
+
+function updateDate(comments){
+    let commentsFinal = [];
+    comments.forEach(comment => {
+        let date = comment.date
+        commentsFinal.push({
+            ...comment._doc,
+            date: date
+        });
+    });
+    return commentsFinal;
 }
 
 module.exports = router;
